requirements.txt
Flask
gunicorn
paramiko
mysql-connector-python
cryptography
werkzeug
python-dotenv